pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "futurnod-researcher"
version = "0.1.0"
description = "AI-powered research API using GPT-Researcher and Tavily API"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "FuturNod Team", email = "info@futurnod.com"}
]
dependencies = [
    "fastapi>=0.108.0",
    "uvicorn[standard]>=0.25.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "gpt-researcher>=0.0.13",
    "nest-asyncio>=1.5.8",
    "httpx>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app", "Agents"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true